add package.json and jest configuration files and renamed StaticPage to Layout
