+++ packages/component-static-pages/config/jest-setup.client.js
@@ -1,0 +1,6 @@
+import Adapter from 'enzyme-adapter-react-16'
+import Enzyme from 'enzyme'
+
+Enzyme.configure({ adapter: new Adapter() })
+
+window.scrollTo = jest.fn()
