+++ packages/component-static-pages/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@elifesciences/component-static-pages",
+  "version": "1.0.0",
+  "main": "index.js",
+  "license": "MIT"
+}
